+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.4.0](https://devt.de///compare/v1.3.3...v1.4.0) (2021-01-09)
+
+
+### Features
+
+* Better pretty printer ([18e42f7](https://devt.de///commit/18e42f771dc386a8a6cfb9329f59c5a246934be0))
+
+### [1.3.3](https://devt.de///compare/v1.3.2...v1.3.3) (2021-01-01)
+
+
+### Bug Fixes
+
+* Not stopping debug server if not running in interactive mode ([4d62c35](https://devt.de///commit/4d62c353384d5cc718b891540abf0e5e344d9326))
+
+### [1.3.2](https://devt.de///compare/v1.3.1...v1.3.2) (2020-12-30)
+
+
+### Bug Fixes
+
+* Adding error display when reloading interpreter state ([f5cc392](https://devt.de///commit/f5cc392a9cd4dd2a1d743fd9f1bc54bb71f1484a))
+
+### [1.3.1](https://devt.de///compare/v1.3.0...v1.3.1) (2020-12-30)
+
+
+### Bug Fixes
+
+* Stopping and starting the processor when loading the initial file. ([981956f](https://devt.de///commit/981956ff9309b8395081d25a2c5711d872818015))
+
+## [1.3.0](https://devt.de///compare/v1.2.0...v1.3.0) (2020-12-29)
+
+
+### Features
+
+* Adding conversion helper for JSON objects ([1050423](https://devt.de///commit/1050423c453169f22da029a52f131e2d3054a1f1))
+
+## [1.2.0](https://devt.de///compare/v1.1.0...v1.2.0) (2020-12-26)
+
+
+### Features
+
+* Adding conversion helper for JSON objects ([0454de3](https://devt.de///commit/0454de30f32b70cb936675c86bf6bdafd4e4bc3b))
+
 ## [1.1.0](https://devt.de///compare/v1.0.4...v1.1.0) (2020-12-13)